[FIX] stock_picking_wave_package_info: stock_picking_wave_fix remove dependency module, and stock.py split file.
